Проект подготовлен к отправке на защиту
