+++ js/image-loader.js
@@ -22,10 +22,10 @@ const successTemplate = document.querySelector('#success')
 
 const closeOverlay = () => {
   closeModal(imgUploadOverlay, body);
-  closeButton.removeEventListener('click', closeButtonListener);
-  document.removeEventListener('keydown', escListener);
-  textHashtagsInput.removeEventListener('keydown', propagationStopper);
-  textDescriptionInput.removeEventListener('keydown', propagationStopper);
+  closeButton.removeEventListener('click', closeButtonClickHandler);
+  document.removeEventListener('keydown', escHandler);
+  textHashtagsInput.removeEventListener('keydown', fieldKeydownHandler);
+  textDescriptionInput.removeEventListener('keydown', fieldKeydownHandler);
   disableEffectPreview();
   disnableScaleChanger();
 };
@@ -41,17 +41,17 @@ const unBlockSubmitButton = () => {
 const showLoadError = () => {
   const message = errorTemplate.cloneNode(true);
   body.appendChild(message);
-  document.addEventListener('keydown', closeByEsc);
+  document.addEventListener('keydown', documentKeydownHandler);
   const closeMessage = () => {
     message.remove();
-    document.removeEventListener('keydown', closeByEsc);
+    document.removeEventListener('keydown', documentKeydownHandler);
   };
   message.addEventListener('click', (evt) => {
     if (evt.target.tagName !== 'DIV' && evt.target.tagName !== 'H2'){
       closeMessage();
     }
   });
-  function closeByEsc(evt) {
+  function documentKeydownHandler(evt) {
     if (evt.key === 'Escape') {
       closeMessage();
     }
@@ -61,33 +61,33 @@ const showLoadError = () => {
 const showLoadSuccess = () => {
   const message = successTemplate.cloneNode(true);
   body.appendChild(message);
-  document.addEventListener('keydown', closeByEsc);
+  document.addEventListener('keydown', documentKeydownHandler);
   const closeMessage = () => {
     message.remove();
-    document.removeEventListener('keydown', closeByEsc);
+    document.removeEventListener('keydown', documentKeydownHandler);
   };
   message.addEventListener('click', (evt) => {
     if (evt.target.tagName !== 'DIV' && evt.target.tagName !== 'H2'){
       closeMessage();
     }
   });
-  function closeByEsc(evt) {
+  function documentKeydownHandler(evt) {
     if (evt.key === 'Escape') {
       closeMessage();
     }
   }
 };
 
-function propagationStopper(evt) {
+function fieldKeydownHandler(evt) {
   evt.stopPropagation();
 }
 
-function closeButtonListener() {
+function closeButtonClickHandler() {
   closeOverlay();
   resetEffect();
 }
 
-function escListener(evt) {
+function escHandler(evt) {
   if (evt.key === 'Escape') {
     closeOverlay();
     resetEffect();
@@ -96,10 +96,10 @@ function escListener(evt) {
 
 const renderImageEditor = () => {
   openModal(imgUploadOverlay, body);
-  textHashtagsInput.addEventListener('keydown', propagationStopper);
-  textDescriptionInput.addEventListener('keydown', propagationStopper);
-  closeButton.addEventListener('click', closeButtonListener);
-  document.addEventListener('keydown', escListener);
+  textHashtagsInput.addEventListener('keydown', fieldKeydownHandler);
+  textDescriptionInput.addEventListener('keydown', fieldKeydownHandler);
+  closeButton.addEventListener('click', closeButtonClickHandler);
+  document.addEventListener('keydown', escHandler);
   enableEffectPreview();
   enableScaleChanger();
 };
